Moved all tpmDirect testing of pcr to pcr_direct_test.go. Added more Tests.
Fixed nits
